+++ platform/discord/bot.go
@@ -332,7 +332,7 @@ func (b *Bot) stop() {
 // AddPlugin takes as argument a plugin and
 // adds it to the bot providing it with the API
 func (b *Bot) AddPlugin(plugin platform.BotPlugin) {
-	plugin.SetAPI(b)
+	plugin.SetAPI(b, nil)
 	b.plugins = append(b.plugins, plugin)
 }