+++ plugin/voteplugin/plugin_test.go
@@ -24,7 +24,8 @@ func TestCreateVotePlugin(t *testing.T) {
 	assert.Equal(nil, p.API)
 
 	api := plugin.MockAPI{}
-	p.SetAPI(&api)
+	storage := plugin.MockStorageAPI{}
+	p.SetAPI(&api, &storage)
 }
 
 func TestVotePlugin_HelpTextAndInvalidCommands(t *testing.T) {
@@ -35,7 +36,8 @@ func TestVotePlugin_HelpTextAndInvalidCommands(t *testing.T) {
 	assert.Equal(nil, p.API)
 
 	api := plugin.MockAPI{}
-	p.SetAPI(&api)
+	storage := plugin.MockStorageAPI{}
+	p.SetAPI(&api, &storage)
 
 	postToPlugin := model.Post{
 		ChannelID: "CHANNEL ID",
@@ -88,7 +90,8 @@ func TestVotePlugin_FailOnPost(t *testing.T) {
 	assert.Equal(nil, p.API)
 
 	api := plugin.MockAPI{}
-	p.SetAPI(&api)
+	storage := plugin.MockStorageAPI{}
+	p.SetAPI(&api, &storage)
 	api.ErrorToReturn = fmt.Errorf("Some error")
 
 	api.PostResponse.PostedMessageIdent.Channel = expectedChannel
@@ -125,7 +128,8 @@ func TestVotePlugin_CreateAndEndSimpleVote(t *testing.T) {
 	assert.Equal(nil, p.API)
 
 	api := plugin.MockAPI{}
-	p.SetAPI(&api)
+	storage := plugin.MockStorageAPI{}
+	p.SetAPI(&api, &storage)
 
 	api.PostResponse.PostedMessageIdent.Channel = expectedChannel
 	api.PostResponse.PostedMessageIdent.ID = expectedMessageID
@@ -210,7 +214,8 @@ func TestVotePlugin_SimpleVoteCounting(t *testing.T) {
 	assert.Equal(nil, p.API)
 
 	api := plugin.MockAPI{}
-	p.SetAPI(&api)
+	storage := plugin.MockStorageAPI{}
+	p.SetAPI(&api, &storage)
 
 	api.PostResponse.PostedMessageIdent.Channel = expectedChannel
 	api.PostResponse.PostedMessageIdent.ID = expectedMessageID
@@ -294,7 +299,8 @@ func TestVotePlugin_DoNotAllowCreationOfTheSameVoteTwice(t *testing.T) {
 	assert.Equal(nil, p.API)
 
 	api := plugin.MockAPI{}
-	p.SetAPI(&api)
+	storage := plugin.MockStorageAPI{}
+	p.SetAPI(&api, &storage)
 
 	api.PostResponse.PostedMessageIdent.Channel = expectedChannel
 	api.PostResponse.PostedMessageIdent.ID = expectedMessageID
@@ -391,7 +397,8 @@ func TestVotePlugin_CreateAndEndCustomVote(t *testing.T) {
 	assert.Equal(nil, p.API)
 
 	api := plugin.MockAPI{}
-	p.SetAPI(&api)
+	storage := plugin.MockStorageAPI{}
+	p.SetAPI(&api, &storage)
 
 	api.PostResponse.PostedMessageIdent.Channel = expectedChannel
 	api.PostResponse.PostedMessageIdent.ID = expectedMessageID
@@ -477,7 +484,8 @@ func TestVotePlugin_CustomVoteCounting(t *testing.T) {
 	assert.Equal(nil, p.API)
 
 	api := plugin.MockAPI{}
-	p.SetAPI(&api)
+	storage := plugin.MockStorageAPI{}
+	p.SetAPI(&api, &storage)
 
 	api.PostResponse.PostedMessageIdent.Channel = expectedChannel
 	api.PostResponse.PostedMessageIdent.ID = expectedMessageID
@@ -569,7 +577,8 @@ func TestVotePlugin_CustomVoteOptionsLimit(t *testing.T) {
 	assert.Equal(nil, p.API)
 
 	api := plugin.MockAPI{}
-	p.SetAPI(&api)
+	storage := plugin.MockStorageAPI{}
+	p.SetAPI(&api, &storage)
 
 	api.PostResponse.PostedMessageIdent.Channel = expectedChannel
 	api.PostResponse.PostedMessageIdent.ID = expectedMessageID