+++ storage/memorybackend/storage.go
@@ -1,0 +1,5 @@
+package memorybackend
+
+type storage map[string]interface{}
+type pluginStorage map[string]storage
+type botStorage map[string]pluginStorage